Easy: Move Gzip Compression to `gzip.{h,cpp}` Files
Summary:
The gzip utilities in `hphp/utils/compression.*` were obviously created when
gzip-based compression was the only compression in HTTP. This is no longer the
case and so it is no longer appropriate to identify these functions as simply
"compression". In particular, I plan to introduce a new compression abstraction
that will in fact abstract over different compression schemes, so this work is
necessary to disambiguate.

Differential Revision: D10023734

fbshipit-source-id: b817706e2bce8525eb7f59f76306bc11792b7e84