Get namespace aliases from parser options
Summary:
In D8804523, I hacked-in the auto namespace map in `Hhbc_id` right before the ID was elaborated, because we couldn't trust that the parser options would correctly flow down from the entry point.

The previous diff (D10235606) inserted the parser options at the entry point, and this diff wires them up to the emitter. In the program emitter and the closure converter we now use passed-in parser options instead of defaults when creating empty namespace envs. This also adds a test for both of those changes.

As of this diff, the parser options are the source of truth for namespace aliases in both the typechecker and compiler.

Reviewed By: oulgen

Differential Revision: D10235645

fbshipit-source-id: 415a1414891895bb15c5a9c15ab9d1d8b42409a9