Fix lookup_foldable_return_type for wrappers with default args
Summary:
When running analyze_func_inline with CollectionOpts::EffectFreeOnly
we would bail at the SetL for a default argument.

Mark side-effect free SetL's effect_free when we're inlining (they're
not effect free when we're not inlining because they affect the local
they're setting).

There are several other opcodes that could benefit from this
treatment, but they don't affect wrapper functions with default args,
so I'll do that as a follow up.

Reviewed By: ricklavoie

Differential Revision: D10411386

fbshipit-source-id: 92e0e02d318a11b7caf01f3276b56163d0298e81