Delete dead code for comparing Objects (as opposed to ObjectDatas).
Summary:
In D10215888, I fixed code that was doing ObjectData <= and >= comparisons in 2 steps, but I missed that we also had similar code on Object: Object::lessEqual and Object::moreEqual. Luckily, they're both dead and can be deleted.
Actually, everything that takes const Object& in comparisons.h is (almost) dead. Fix the 2 call sites, and =delete the dead code.
This deletion makes the comparison helpers on Object and Object::toInt64ForCompare and Object::toDoubleForCompare dead, as well as the comparison methods on Array that take const Object&, as well as the check_collection_compare helpers in builtin-functions. Delet them all.

Reviewed By: ricklavoie

Differential Revision: D10423317

fbshipit-source-id: db9ab57f31ecb8fbe9a52963f6676282ec3915ba