Improve typing of array assignment
Summary: This diff improves typing of assignments to array elements in a way similar to how D9082246 improves typing of array appends.  Namely, we allow assignments to collections with copy-on-write semantics of elements of incompatible types.  We also handle the case of collections of unresolved types properly (assignment to such a collection should only be allowed if assignments to each of the types comprising the unresolved type are valid).  Finally, we allow assignments to tuple elements.  See the new test cases exemplifying the improved behavior.

Reviewed By: CatherineGasnier

Differential Revision: D10258204

fbshipit-source-id: 9261a46e8e09d888a5442f76f54002e5325ba974