Add verification to FPushCtor for classes with/without reified generics
Summary:
This diff adds verification to `newInstance` and `newInstanceReified` which are both used by the interpreter and the JIt either directly through `Fpushctor` or through `AllocObj` or `AllocObjReified`.

This verification checks whether reified classes are reified with correct number of reified generics and making sure non reified classes are not passed in reified generics. It adds more user friendly error messages to explicitly indicate what the mistake is.

It also cleans up MaybeReified flavors in this verification process as we no longer need it since FPushCtor tells us whether we expecting reification or not.

This also results in removal of TSOrTrash type which was very unsafe.

Reviewed By: ricklavoie

Differential Revision: D10430295

fbshipit-source-id: bc082774f496b33186a60e45d04192225850b000