Fix the jit's understanding of when VerifyParamFail may coerce.
Summary:
The jit and the implementation of VerifyParamFail were out of sync on when it could coerce the param. If it coerced but the jit thought it didn't, and you used the param in the same tracelet that VerifyParamFail is in, then you would get corrupt data or a crash.
Fix this by adding a verify_fail_may_coerce helper function. In the jit, pessimize the type of the param in the frame state when it returns true. In TypeConstraint::verifyFail, assert that verify_fail_may_coerce is true before performign any coercions.
Add a unit test that would crash before this fix (fun thought it had an object but due to coercion actually had a string).

Reviewed By: ricklavoie

Differential Revision: D10458575

fbshipit-source-id: bd96bbe0546b3638dc8446a95cfe1e5059c07c6b