Visit shape keys in AAST visitors
Summary:
It looks like visitation of shape keys was overlooked in D6481759--these base visitors currently only visit ShapeMap values. After this change, we will visit both keys and values.

The `endo` visitor is meant to be equivalent to `map`, but should avoid allocating new nodes when unnecessary. We don't use `endo` visitors much, and the existing implementation already reallocated unconditionally, so I haven't taken the time to write a proper `endo` implementation and called out the problem with a FIXME instead.

Reviewed By: arxanas

Differential Revision: D10470531

fbshipit-source-id: 596cdaab71b46abb361eb85f6aef48a43f22e45e