Cleanup ConcatPeephole
Summary:
I've never felt confident about its various checks.

eg before processing each opcode it checks the stack depth to see if
the previous instruction should terminate the concat stream. That
seemed suspect because the previous instruction might have popped 10
elements, and pushed 10 elements - which would certainly need to
terminate any concats operating on those elements (this did work,
because the operator stored in srcStack for those elements would no
longer be a Concat - but the checks for that were not centralized
making it hard to follow).

Switch to a more stratightfoward algorithm that terminates a concat
stream if the last thing that wrote to its target cell wasn't a
concat.

I also noticed that it never merged streams, meaning that it would
miss things like

  CGetL $a
  CGetL $b
  Concat  # <- starts a stream
  CGetL $c
  CGetL $d
  Concat # <- starts a second stream
  Concat # <- adds to second stream

After this change, the last concat will also merge the second stream
into the first, and then add itself to it.

Reviewed By: ricklavoie

Differential Revision: D10466864

fbshipit-source-id: b817551e50897c6b05add9b02cf4592084c77584