Refactor Response Compression into Separate Class (v2)
Summary:
This diff pulls the logic to compress responses out of the `Transport` class
and into its own class, the `ResponseCompressor`. Because compression was
very tightly to the transport layer, this is a fairly invasive change.

The motivation for this change is to set up HHVM to support stateful and/or
dictionary-based compression schemes. Such a scheme will use additional
headers on the request and response and at least initially will not be in the
open source HHVM (so we will need to delegate to a dynamically linked module).
`Transport` therefore needs to be able to call out to something else to make
these decisions.

Depends on D10316086.

Supersedes D10022047.

Differential Revision: D10316138

fbshipit-source-id: e20d2e1fcea4645f04e188ba3e001b395d90ed9c