Fix a stack overflow in GetScalarArray
Summary:
hphp/test/slow/parser/memory_exhaust.php was expected to fail in the
parser, or overflow the stack while deserializing the string
representation of the array. In gcc-8 lto builds, it would
successfully deserialize the string, but then crash with a stack
overflow in GetScalarArray.

This adds the necessary check to GetScalarArray to prevent a crash,
and makes the test pass.

Reviewed By: alexeyt

Differential Revision: D10481545

fbshipit-source-id: 6e40d1335e108dc2f3362e1396b8ec92a965f574