Distinguish between client -> server and server -> client requests in integration script
Summary:
The Hack LSP integration tests don't distinguish between client -> server and server -> client requests and hence can interpret a server -> client request as a response to a client -> server request if the id's match up (the JSON shape determines if something is a request or response).

This diff distinguishes between the two so that I can implement the showStatus request and have it pass integration tests.

Reviewed By: ljw1004

Differential Revision: D10472854

fbshipit-source-id: 6d2782d05a9ed22a2dddab787ed6a202ad8d4b12