Fix WatchmanEventWatcher to prevent starting hh_server in hg.filemerge state
Summary:
`WatchmanEventWatcher` is a daemon that is intended to have a lifespan longer than that of single server, or single server monitor, which allows it to more reliably answer the question “are we in the middle of rebase”.

It has a bug (?) where instead of only looking at `hg.update` state enter/leave events, it just treats any leave as end of update. This ends up including things nested inside `hg.update`, which makes it incorrect, with the worst offender being `hg.filemerge` being entered/left for every file with merge conflicts WHILE the rest of files are still mid-update.

There was already test that is failing for similar reason related to `hg.transaction` events, but it was disabled as flaky long time ago ¯\_(ツ)_/¯.

Reviewed By: 2BitSalute

Differential Revision: D10511311

fbshipit-source-id: 5e3c26fdf119ee41df4322db15a457c7a4315493