Const-fold (v)sprintf.
Summary:
Make it safe to fold (v)sprintf by checking getThrowAllErrors and raising a notice if we're about to rely on locale data.
Mark sprintf and vsprintf foldable.
While I was here, I noticed that sprintf was creating a StringBuffer that's one byte over a size class; fix the size.

Reviewed By: jano

Differential Revision: D10510267

fbshipit-source-id: 8a89501b70c0fc92595a1cd51a5c241acda32c1b