Fix callstub clobbers
Summary:
Nothing prevents unique-stubs from using caller or callee saved
registers. However the clobber set for the callstub vasm instruction explicitly
did not include the callee saved registers. This meant that a unique stub could
clobber a register that is expected to be kept alive across the call. We seem to
get away with this just by chance. Fix this by explicitly unioning in all the
registers that a stub is allowed to use into the clobber set (this is a bit more
precise than just including the callee saved registers).

However this introduces more register pressure on the caller, which is a problem
for prologues. Prologues use the same ABI as unique-stubs and thus cannot
spill. Increasing the clobber set was enough to cause attempting spilling in
prologues, leading to asserts. Fix this by slightly shrinking the set of
registers that unique-stubs are allowed to use (this shrinks the clobber set,
reducing the register pressure on the caller). This requires that we have
prologues use its own unique ABI since it defeats the purpose to also shrink the
registers that it can use.

I didn't change the register sets for ARM or PPC because I cannot test if its
sufficient or even necessary.

Reviewed By: mxw

Differential Revision: D10518285

fbshipit-source-id: e45ce5de170c0e4fdedf8465c7ce959560bc12f4