Move base class 86reifiedinit methods to runtime
Summary:
This diff moves the emitting of 86reifiedinit methods to the JIT. When AttrNoReifiedInit is not set on a base class, it emits a fresh 86reifiedinit method. Initially, this attribute is not set on all classes but HHBBC sets it on classes that do not have any reified classes that extends the base class.

The runtime performance should be identical but HHBBC should run much faster now.

Reviewed By: markw65

Differential Revision: D10504399

fbshipit-source-id: 27898664e19c1cac414e60c765b5e4efd8fd1e5a