Fix an objprof crash
Summary:
Inside APCObject::createObject, we can end up executing php code while
an object's properties are not yet initialized (ie contain garbage,
not KindOfUninit). This is normally ok, because the gc is tolerant of
garbage pointers, and normal php code can't see these objects. But
because ObjProf walks the heap, it *can* find them, and crashes when
it tries to inspect them.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D10561289

fbshipit-source-id: 6444c847a14e178cc218e3d9ec5a9e045e1059ce