Reduce cost of creating constant arrays
Summary:
We were still creating a new ArrayData for each Add*ElemC, resulting
in O(n^2) time complexity.

This attempts to just incrementally build the array in the common case
that each new Add*ElemC does in fact extend the array.

This seems to work for opt builds, and gets a significant speedup
(2-3x) in dbgo builds; but dbgo builds are still O(n^2) because of the
checkInvariants call in Type::Type and Type::~Type.

Reviewed By: ricklavoie

Differential Revision: D10500760

fbshipit-source-id: 1bcb607ddcd3ee3d22209bfc092710634f9314e2