introduce `apc_size` function for fetching in-memory size of apc data
Summary:
There are certain data structures that take up a large percentage of APC memory in production and other environments and occasionally it's worth optimizing their size in memory. There is currently a way to figure out how big a structure is in memory, but it's quite roundabout, requires a long-lived server to run the admin port on, and requires access to the admin port in the first place.

In order to improve this process, `apc_size` will report the same size as the admin dump command for `inmem` keys (reported from `APCHandle`) and will unserialize primed keys into an APC handle and update the store to report their size. This keeps the size the same whether or not a key is primed or not or it's been accessed before; these details aren't worth leaking and this is mostly for local debugging and iteration anyway. Primed keys will likely never go through this method.

Reviewed By: alexeyt

Differential Revision: D10561091

fbshipit-source-id: 18911ae50792c27f1c149575547aed7aa2255172