Be more explicit about the default case in simplify_subtype
Summary: Replace the final `_, _ -> default ()` case in `simplify_subtype` with a more detailed case analysis.  I leaned on the compiler to suggest the missing patterns, and initially mapped everything to `default ()`.  Many of the cases could be refined, and some of the patterns could be merged.  I didn't attempt any more substantial reshuffling of the code (similar to D10447650) to simplify reviewing.  Will continue to clean things up in a subsequent diff.

Reviewed By: CatherineGasnier

Differential Revision: D10851913

fbshipit-source-id: 4e2f148b8657217da564b3a2851f1fa528cecbc7