Ban goto from the typechecker
Summary:
Being able to jump to arbitrary locations in code is pretty wild and makes typechecking more complicated. We should not have this feature in Hack. However, it will remain in the AST for language features that may require finer control flow when desugaring.

While in the neighborhood, moved some loose `goto` tests into the already-existing `goto/` directory.

Reviewed By: fredemmott, michaeltingley

Differential Revision: D10476126

fbshipit-source-id: 850b1a9e027919fabeb85fb27ed2cfa80f909ecc