Type deserialization infrastructure
Summary:
Base commit for the ability to deserialize types that we've previously deserialized.

The actual implementation will be split up into future commits.

For now, I'm only checking `locl` types (such as those types that are inferred in a function's body). I may add checking `decl` types in the future as well.

Reviewed By: ljw1004

Differential Revision: D10044186

fbshipit-source-id: 12c7b4baf6eb8f3848fe6000622d3ac1f69b661c