Avoid cases where server is restarted in Died_config_changed state (leading to lack of saved state due to version mismatch)
Summary:
A story out of OOM logs - 2 out of 10 servers I checked OOMed because they had no saved state, because they were running wrong version. This is due to a bug - `Died_config_changed` is a special sate that should only be exited by new client connecting (assuming that new client exercised all the version lookup logic, which will cause it to kill the server if version really doesn't match).

All the auto restarting logic was disregarding that and restarting the server anyway.

Reviewed By: ljw1004

Differential Revision: D12813597

fbshipit-source-id: 9ca906bd2752653436ab7be1611917a0755b7fa9