disable intercept surprise flag when all intercepts are cleared
Summary:
The intercept surprise flag comes at a huge runtime cost, since we need to check the surprise flag in all our function entry and exit points.

The API allows for clearing an individual handler by passing in something that's boolean false (`null` is a common value) for the callback. We currently set the flag when we first set a handler (either a named one or a global intercept handler), but never clear it out when removing the final one.

`fb_intercept` will always be sketchy for performance, but if it's being used in a guard, we can at least recover the performance outside of that guard.

Reviewed By: markw65

Differential Revision: D12824240

fbshipit-source-id: 023466d1d86d1a146136a310f068a1c2b93a6884