add clang-specific EXTERNALLY_VISIBLE macro defn, plus VISIBILITY_DEFAULT
Summary:
Clang doesn't know the `__externally_visible__` attribute; it seems that's a gcc-specific thing.

For clang, let this macro expand to two attributes; `used` and `visibility(default)`.  The `used` attribute convinces Clang that the symbol needs to be emitted, and `visibility(default)` is in case something is set (by command-line flag, perhaps) to hidden visibility.

Reviewed By: mgrice

Differential Revision: D10493483

fbshipit-source-id: aa09c42601ee3474fffe5d30a7beed4c76460411