Add Execution_operator to the annotated AST
Summary:
In order to compile from an annotated AST, we need to support the execution operator (backticks).

This is code of the form:

    $x = `ls -l {$directory}`;

This syntax is forbidden in Hack, but valid PHP.

Add this variant to the annotated AST types and update the visitors. Throw an error if the type checker is ever asked to type check this (it shouldn't).

For the abstract interpretation logic, treat the execution operator the same way we handle other string expressions.

Reviewed By: andrewjkennedy

Differential Revision: D10868739

fbshipit-source-id: fd16808edd6e51c1d0b82f4812fe843e797025e0