Check first for likely child state when awaiting
Summary:
When awaiting a child coming from a FCall with async eager offset, the
child is unlikely to be finished. Instead of checking first for success and
then for failure, check directly for the likely case of being unfinished.

Also, fix a bug in ifThenElse where an empty unreachable block was created,
just to be considered as reachable and mergeable by the outer ifThenElse.

Reviewed By: paulbiss

Differential Revision: D10850606

fbshipit-source-id: ed1a2261f9bd48886a2a464f4c51f161ab50b9a1