Combine hashHelper_NvGetStr and NvGetStr
Summary:
Combine hashHelper_NvGetStr* and NvGetStr* into a single function.

These functions are written in assembly, and BOLT views NvGetStr*
as a function with multiple entries which prevents it from
changing the code layout. With this diff, 2 functions are combined
into one, and BOLT can re-arrange the code based on a profile data.

Reviewed By: binliu19, ottoni

Differential Revision: D12869426

fbshipit-source-id: 42f9d378b9b7221726d68b26cd23d52b0ea6d115