Log to Scuba when failing to save or load declarations in saved state
Summary:
Currently, when something goes wrong while saving or loading declarations, we silently continue saving or loading the saved state. Saved states are still useful without class declarations, so there is no need (yet) to completely give up because there is an issue with this new feature. However, as we roll it out, we need signal if it isn't working. This logging is intended to provide that signal.

Once this feature has been rolled out and demonstrated to work, we can fail hard instead of simply logging.

Reviewed By: ljw1004

Differential Revision: D10864988

fbshipit-source-id: 0ced203dbc4d67ddb5db0c42341470966bfce6d7