Option to nerf autocomplete / docblock features to reduce FFP callsand see if it helps
Summary:
Logging added in D10376916 reveals that autocomplete and docblock finder (effectively part of autocomplete) cause disproportionately many files to be parsed[1], especially for commands that are on a critical “after keystroke” path.

Before trying to optimize them, I'd like to test whether it will do anything to our metrics. To test this cheaply, I add an option that disables features causing all this parsing and I'll turn it on for a few % of unlucky users for a day or two. The experience will be degraded in few ways:

* we'll suggest global names in some ways they don't make sense and what we avoid now: abstract classes and enums after new, non-traits after “use”; basically downgrading it to a prefix match on all class names
* no icon difference between classes, traits, enums and interfaces
* return type information on functions will be missing
* some non basic doc blocks will be missing

To reiterate, the intention is not to remove those features, but briefly test whether it's worth spending time optimizing them, so don't worry about how bad it sounds.

Reviewed By: pittsw

Differential Revision: D12879624

fbshipit-source-id: 0856d51105bfe68f521a7fde330be18de2c29e14