Option to utilize ClientProvider.sleep_and_check time to perform Gc.major_slice
Summary:
All the things I tested in {D10456676} did almost nothing to prevent the 1+ second pauses in master processes causing us unhealthy IDE minutes. We seem to be just too slow no matter what, and big source of slowness is having to GC all the garbage we allocate.

With this diff I'm going to test the hypothesis that maybe there are enough breaks between different commands that now are wasted, but could be utilized to reorder the heap and smooth out the peaks.

I don't really believe that it will work on it's own (we send too many too expensive commands in a row on single keystroke), but maybe it will help when combined with some other things (like D12879624).

Reviewed By: pittsw

Differential Revision: D12879607

fbshipit-source-id: 6000fda28c209455d35bacee51d112e9a10a4921