Make Tmixed a decl-only type
Summary: `Tmixed` is desugared into `Toption (_, Tnonnull)` by `Typing_phase.localize_with_env` quite early in the pipeline, and we essentially assume that we never encounter `Tmixed`, but because it's currently defined as a type that can exist in both `decl` and `locl` phases, we still have to include `Tmixed` in pattern matches.  Let us make `Tmixed` a `decl`-only constructor and thereby statically ensure that in cannot occur as a `locl`-phase type.

Reviewed By: andrewjkennedy

Differential Revision: D12857484

fbshipit-source-id: 1515718fe880e42d2563b95d7b045989eaab67bf