Fix assertion in CurlResource::close()
Summary:
The Curl extension tries to proxy exceptions raised from the various Hack
callbacks that it has registered. To avoid throwing thru C code (which is
probably not safe), it catches any exception raised, stores it, and then when
the C callbacks are completed, it will rethrow the exception. There's a number
of assertions to verify that there's not a queued exception outside of this
callback context.

However, the code to process any queued exceptions is itself susceptible to
throwing exceptions. If this happens, the queued exceptions will be left behind
and we'll assert later. Fix this by using SCOPE_EXIT to ensure that all the
resources have their queued exceptions released, even if we throw.

Reviewed By: markw65

Differential Revision: D10853978

fbshipit-source-id: 726936681a22d39b9dcc8863e168b437eaea9981