Move misc THREAD_LOCALs to RDS_LOCALs
Summary:
This moves several thread singleton objects to be stored in RDS.  This shows
that migrating is possible without huge performance blow backs.  This change
is perf neutral.

Of particular interest is the iter_table thread local.  We have extremely hot
code checking the existence of mutable iterators.  This required using the
RDS_LOCAL_HOTVALUE api to achieve parity.

Depends on D10494926

Reviewed By: arnabde03

Differential Revision: D9583875

fbshipit-source-id: 200d9f4fa299d1e8a31b0fe236afac45916cf649