Test saving decls in incremental saved state
Summary: When exporting class declarations to be stored in the saved state, if a declaration is not present in the decl heap, Decl_export will declare the class if it exists. These test cases exercise this behavior (which only arises when generating incremental saved states, since all declarations should be present when generating a saved state after a full init).

Reviewed By: dabek

Differential Revision: D10864984

fbshipit-source-id: eb2fd247bb55bc2e2f0d9a3e06fab4de861b7d4b