Test loading declaration of hot similar class with decl FIXME
Summary:
If we fail to oldify declarations we loaded from the saved state which are in similar files (that is, files whose decl AST hash has not changed), positions in these declarations could be stale. When we recheck the file (since we always recheck similar files), we will re-parse it, updating the DECL_HH_FIXMES heap with new positional information (which no longer corresponds to the positions we loaded from the saved state). This can cause us to fail to suppress the error message (or alternatively, to report the wrong positions in error messages involving a stale loaded declaration).

We need to make sure the loaded declaration is oldified even though the decl AST hash has not changed. This test ensures that we do.

Reviewed By: dabek

Differential Revision: D12875742

fbshipit-source-id: b7ba5ddb87f630df34d8dfee17b11bbe676ff459