Test loading declaration of hot unchanged interface with decl FIXME
Summary:
When we load the declaration of a hot class from the saved state, and the file containing the hot class has not been changed in master or locally since the state was saved, we end up with no information about the FIXMEs in that file in memory. If we don't re-parse the class at some point before its loaded declaration is used, we can fail to suppress errors in it which have suppression comments.

This test ensures that we successfully suppress such an error in one of these scenarios.

Reviewed By: dabek

Differential Revision: D12877486

fbshipit-source-id: 41473a0d3dc5826597c018e34e4910c9e96ce36f