Don't form optimized regions to cover blocks with low profile weight and truly enable HotCFG
Summary:
This diff enables the HotCFG region selector to truly prune blocks that have a
low profile weight relative to the entry of the region (the thresholds were set
to 0, so HotCFG was doing the same as WholeCFG).  This diff also extends adds
logic to avoid creating a region just to cover a block that has low profile
weight relative to the main function entry.  The idea here is to avoid creating
cold regions that likely to be large.

Reviewed By: mofarrell

Differential Revision: D12897592

fbshipit-source-id: 9bce7e817d4d0936d70f9e96864aec3ce745b0cd