Use DecRefProfile for DecRefNZ
Summary:
A DecRef may be turned into a DecRefNZ during refcount optimizations, and this
is the only way DecRefNZs are produced.  However, DecRefNZs were using
RefcountProfile to try to improve the code layout, which wouldn't find any
profile data (since we don't do refcount opts on profiling translations).  This
diff switches DecRefNZs to use the same DecRefProfile used for DecRefs, allowing
them to be properly optimized based on profile data.  This also requires adding
the extra data used for DecRefs to DecRefNZs, which encodes the local ID used to
make the profile data more accurate.

Reviewed By: mofarrell

Differential Revision: D12833240

fbshipit-source-id: 2e02f1ae92e89f38aebe39c03407cb168364dbe8