Simplify shift/test -> shift
Summary:
After every FCall;Await sequence where the call might be to an async
function, we test the returned value to determine if the call returned
eagerly.

The test sequence currently shifts rdx right by 32 bits, and then
tests to see if the result is zero. The shift already sets the flags
though, so we can skip the test, so this adds a simplification to kill
it where appropriate.

Reviewed By: ricklavoie, jano

Differential Revision: D12871391

fbshipit-source-id: cbd6f49beb2ca4cf72dd9d79c70b3896cbd85eaf