Optimize and->test sequences
Summary:
There was already an attempt to turn "and; test" into "test", but that
only works if there are no more uses of the and's output.

Instead, we always convert "and; test" to "and", and add a new
optimization to convert and to test if its output is unused.

I was seeing a lot of

  andi 0xf, $eax
  test $rax, $rax
  jz ...

in async code, where $rax is still live, so the original optimization didn't catch it. Now the test is gone.

Reviewed By: jano

Differential Revision: D12896181

fbshipit-source-id: 02ea277d4da94620994d8833461fd0cff9b1d328