Don't emit DefCls for anonymous classes
Summary:
For anonymous classes we'd emit

  DefCls <N>
  FPushCtorI <N>

FPushCtorI will define the class if its not already defined, so the
DefCls is redundant.

There was a bug however; in jitted code, we'd emit an interpOneDefCls
for the DefCls, and then conditionally call Unit::defClass if the
class was not already loaded. The Unit::defClass path was never tested
because of the interpOneDefCls, and was missing a VMRegAnchor. This
adds one.

Reviewed By: ricklavoie

Differential Revision: D12839721

fbshipit-source-id: 89ccb8a56fca9b71b634f7063ad9697c8af0c810