Remove <static> from AKdependent and use this instead
Summary:
The type of `$this` and the type denoted by `static` in constructs such as `new static` and `static::class` was represented as a nested use of `AKdependent`, displayed `<static> as this as t`. This is probably historical legacy as the type-checker doesn't distinguish between the use of the "static" and "this" constructors in `AKdependent`. We can remove the use of "static" entirely, to simplify the code.

Also there are a number of places where we call `Typing_dependent_type.make` that simply (check the flow!) return the type unchanged. We can remove these calls.

Reviewed By: manzyuk

Differential Revision: D12916969

fbshipit-source-id: 662b01125359583495512ebf247429aeadf81546